Explicit windows-2019 instead of windows-latest (https://redirect.github.com/duckdb/duckdb/pull/13883)
Using windows-latest meant defaulting to a recent Windows SDK, that
caused issues such as https://github.com/duckdb/duckdb/issues/13848 or
https://github.com/duckdb/duckdb/issues/13864

Issue is, at his core, that we were depending on a Microsoft Visual C++
Redistributable package that is not installed by default on Windows
2019.

Thanks everyone for reporting this, and @szarnyasg and @Mytherin for
helping figuring this out.

This PR provide a way forward for build artifact that will be built, but
also allow a blueprint (and the means) on how to rebuild `v1.1.0`
release Windows artifacts, that will likely happen tomorrow. This will
offer a solution after downloading new artifacts, after we can confirm
they work.

Other duckdb repositories might have the same problem, and would need to
be reviewed.

Upstream-commit: https://github.com/duckdb/duckdb/commit/e2b177b759dbb7cabae0c0afd041bb7de2a9e698