Copy From Database - create a balanced UNION ALL tree instead of a depth-first union all tree to avoid stack overflows (https://redirect.github.com/duckdb/duckdb/pull/13889)
`COPY FROM DATABASE` creates a number of `INSERT` statements, unified
through `UNION ALL` operators to execute this as a single plan. This PR
reworks how these union all statements are created to create a balanced
tree instead of a depth-first tree, leading to a much less deep tree.
This avoids stack overflows later on when copying databases with many
tables.

Upstream-commit: https://github.com/duckdb/duckdb/commit/081a748340c4fcd3b3652230a02432afae72bbb3