Fix data race when writing GeoParquet (https://redirect.github.com/duckdb/duckdb/pull/13962)
When writing GeoParquet, we stored the column-level auxiliary geoparquet
statistics/metadata directly in the `GeometryColumnWriter`, which is not
thread safe, causing issues when multiple row-groups are written
concurrently.

This PR moves the column-level geoparquet metadata from the
`ColumnWriter` into the associated `ColumnWriterState`, and also
introduces a write-lock on the file-level geoparquet metadata struct so
that the threads don't race each other when flushing the stats after
finishing a row group.

Additionally I've changed the `GeometryColumnWriter` into a
`WKBColumnWriter` that directly subclasses the `StringColumnWriter`
instead of trying to template the base class as that is significantly
messier to do when we also need to wrap the associated state. This makes
the code a bit cleaner, and makes sense given that the WKB geometry
encoding is the only one we support right now anyway. I've also realized
that the implementation for the other encodings is going to end up
different enough that there is probably no point in trying to generalize
all the geoparquet specific encoding logic into a common base class
anyway.

Closes: https://redirect.github.com/duckdb/duckdb/pull/13914
Upstream-commit: https://github.com/duckdb/duckdb/commit/b8c5fa937919631b759a70e33f068aa05de8bd36