Check vector type in GetVectorScanType to avoid concurrent race when updating validity (https://redirect.github.com/duckdb/duckdb/pull/13965)
When scanning a column, `GetVectorScanType` needs to be consistent
between the validity column data and the top-level column data,
otherwise an `InternalException` is thrown. As the return value of
`GetVectorScanType` changes based on whether or not updates are found,
this could cause an issue if an update was performed in between the scan
of the top-level column data and the validity column data. This PR
resolves that by checking if the result vector type was already set to a
non-flat vector.

Upstream-commit: https://github.com/duckdb/duckdb/commit/35a104529b56c4f4f1e383e2ead26d6047d3442e