fix: remove http prefix from proxy value when present (https://redirect.github.com/duckdb/duckdb/pull/13973)
DuckDB currently throws the following
[error](https://github.com/duckdb/duckdb/blob/main/src/common/http_util.cpp#L21)
for http_proxy values formatted as `http://<proxy_fqdn>:<proxy_port>`:

`InvalidInputException("Failed to parse http_proxy '%s' into a host and
port", proxy_value)`

This PR creates a copy of the `proxy_value` argument within
`HTTPUtil::ParseHTTPProxyHost` and removes the `http://` prefix if it is
present before parsing the hostname and port. It also includes a unit
test to cover this new condition.

Upstream-commit: https://github.com/duckdb/duckdb/commit/86723c9912fde7b76d3863b2ccd2d4333251c4af