Fix partitions on wide tables (https://redirect.github.com/duckdb/duckdb/pull/13988)
On main branch, `Tables.partitions` fails on table with more than 10
columns. The (somewhat convoluted) reason is that julia `ntuple`
function when fed a `UInt64` as second argument, casts it to `Int` only
if it is `<= 10` (see https://github.com/JuliaLang/julia/issues/55790).
This works in our favor for small `n` (number of columns), as the
function `ColumnConversionData` expects a `Int`, not a `UInt64`, but for
larger `n` attempting to collect a `Tables.partitions` fails.

The first commits is a minimal fix + test, whereas the second commit
also adds a small refactor to simplify code and make it more robust to
this.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6b7d3ec9adbc10818eef760490f3966f08338e2d