fix julia partitions for streaming result (https://redirect.github.com/duckdb/duckdb/pull/14000)
Previously, calling `nextDataChunk` on a streaming result failed to
update the `chunk_index`. This in turn caused the check to prevent
double iteration on `Tables.partitions(q::QueryResult)` to fail for
streaming data.

Now the `chunk_index` is always updated after a successful
`nextDataChunk` call. I've added tests that iterating partitions works
when it should and throws `NotImplementedException` if one attempts to
iterate twice.

Upstream-commit: https://github.com/duckdb/duckdb/commit/5a9ea8231c8831cf3a4e9940bbcbc4dc273d1528