Fix crash in the shell caused by printing blobs that failed to cast (https://redirect.github.com/duckdb/duckdb/pull/13983)
When using the (undocumented) `.dump` command in the shell, while
opening a database that contains blob-aliased types provided by an
extension that is unloaded, the CLI would crash as the shell would
detect a value to be present, but the the unknown type would fail to
cast to BLOB (as the extension, and thus any cast functions are not
available) and return a nullptr.

I've modified the shell and the sqlite3 api wrapper to handle this case,
but I'm not sure if this is the best fix or what the intended behavior
should be when interacting with unknown types from the sqlite api. If we
were to fall-back into a reinterpret cast, im not sure that's desirable
when casting within the `sqlite3_column_has_value` as the type check
then become essentially meaningless. Also there is no guarantee that the
extension in question actually provides a valid reverse cast from the
underlying type (e.g. BLOB) to the extension type even when loaded. It
feels like the only sane thing is to either error out or turn all
unknown types to NULL.

Closes https://github.com/duckdb/duckdb_spatial/issues/391

Upstream-commit: https://github.com/duckdb/duckdb/commit/9daac6a0b79611ed29bca37b45534caab288afca