Less salt (https://redirect.github.com/duckdb/duckdb/pull/14173)
Apparently ARM64v8 uses 52, not 48 bits for the pointer (not all
versions of ARM64v8, but just some). We could also `#ifdef` our way
around this but this seems fine. It reduces the effectiveness of the
salt by 16x sadly, but it's still reduces the chance of having to chase
the pointer even when keys are different by 4096x (was 65536x).

Fixes https://github.com/duckdb/duckdb/issues/14140

Upstream-commit: https://github.com/duckdb/duckdb/commit/92c65a4341c57f313dbeba5acc7b1fb917808010