Avoid schema changes with IF NOT EXISTS (https://redirect.github.com/duckdb/duckdb/pull/14143)
We've noticed that repeated calls of `CREATE TABLE IF NOT EXISTS` for
the same table were causing catalog versions to go up (causing rebinds
for other queries). As ORMs are frequently using this pattern to ensure
that a table / schema exists before inserting data (i.e. packaging the
CREATE TABLE IF NOT EXISTS into the same TX as the INSERT statement),
this led to rebinds happening all the time.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1eac05ecd3a6b8ec2cdf0c53ccece7ca2effef26