Add option to ignore GeoParquet, disable spatial autoloading when reading GeoParquet (https://redirect.github.com/duckdb/duckdb/pull/14297)
Closes https://github.com/duckdb/duckdb-node/issues/124
Closes https://github.com/duckdb/duckdb_spatial/issues/423

This PR changes the GeoParquet reading behavior so that the spatial
extension is no longer _auto-loaded_ when GeoParquet metadata is
detected. Spatial was not marked as autoloadable and the way the load
was issued did not respect auto loading settings either. If spatial
would become auto-loadable in the future we can change this behavior
back (but this time also respect auto loading settings).

This PR also adds a new setting to the parquet extension
`enable_geoparquet_conversion = true (default)` which controls whether
geometry data will be converted when reading/writing GeoParquet files
(and the spatial extension is loaded). This enables you to "fall back"
to reading writing parquet files normally even if they have geoparquet
metadata. This is primarily useful when trying to read a geoparquet file
that is incompatible with the current geoparquet support in duckdb,
either because it uses an unsupported version/encoding or because it is
non-standards compliant (or the metadata is broken in some other way).

Additionally, I've also moved over the geoparquet tests and test files
into this repository for now. Since they `require spatial` they won't be
invoked during our normal CI runs (I think?), but it still useful when
testing locally.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6dc2e9375870e60f82becb1cece4cc878289d3b8