Use table-level locking when acquiring shared locks (https://redirect.github.com/duckdb/duckdb/pull/14370)
This is a follow-up to https://github.com/duckdb/duckdb/pull/13712,
using a more fine-granular locking approach.

https://redirect.github.com/duckdb/duckdb/pull/13712 has fixed an issue where checkpoints together with concurrent
queries that referenced the same table multiple times could lead to
deadlocks. The locking approach in that PR used a transaction-level lock
to coordinate lock acquisition for the same table.

In case of a query having many table scans (for different tables), this
could potentially now block all executor threads if only a single table
lock could not be acquired (e.g. due to a concurrent checkpoint). One
thread would be blocked trying to acquire the checkpoint lock for the
given table, and many others on the active_locks_lock (even though they
would only be interested in acquiring the checkpoint lock for another
table that wasn't being checkpointed).

Upstream-commit: https://github.com/duckdb/duckdb/commit/4756244efa04d204be6f20d55036fc503b7ed49c