add index plan callback to IndexType (https://redirect.github.com/duckdb/duckdb/pull/14511)
This PR adds another hook to the `IndexType` class to allow indexes to
control how the physical plan gets generated from a logical `CREATE
INDEX` plan.

Previously the `CreatePlan` for the `LogicalCreateIndex` operator was
hard-coded to only plan `ART` indexes. Custom index types (such as those
in vss and spatial) relies on optimizer extensions to "hijack" the query
plan and replace the `LogicalCreateIndex` with e.g.
`LogicalCreateHNSWIndex` before physical planning could begin. This hack
and resulted in a lot of duplicated and very advanced code in these
extensions, and also came with the unfortunate side effect that you
could not create these index types at all if the optimizer was disabled.

This is just the first step in a larger extension index rework Im
working on, and I want to make the interface here even tighter in the
future by e.g. handling sorting/null filtering/expression type
validation before we hand of control to the extension, as I think that
is something that could be generalized and/or is interesting for most
index types and is a bit complicated to do right now.

Upstream-commit: https://github.com/duckdb/duckdb/commit/78b65d4a9aa80c4be4efcdd29fadd6f0c893f1ce