CLI: Add support for `.thousand_sep` and `.decimal_sep`  (https://redirect.github.com/duckdb/duckdb/pull/14721)
This PR adds support for the `.decimal_sep` and `.thousand_sep` options
that can be used in the CLI to configure how these are rendered. Note
that this is only respected in `duckbox` mode currently. Example usage:

```sql
.decimal_sep ,
.thousand_sep _
SELECT 12345.6;
┌──────────────┐
│   12345.6    │
│ decimal(6,1) │
├──────────────┤
│     12_345,6 │
└──────────────┘
```

Note that there is one restriction - this currently does not work with
nested types yet but only for numbers at the top-level.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9e446321cdc1482c9750654d6bb4e1a3ac12ed4f