adaptive filters should not reorder filters that can throw (https://redirect.github.com/duckdb/duckdb/pull/14672)
fixes https://github.com/duckdblabs/duckdb-internal/issues/3180

And this is a followup to https://github.com/duckdb/duckdb/pull/13617 

If the table being filtered is large enough, the order of the
expressions will get moved around due to adaptive filtering. To prevent
this, expression reordering now places severe penalties on expressions
that can throw, effectively forcing them to be executed last. Then
during adaptive filters, we set the likelihood of swapping filters that
can throw to 0. This allows other filters to be swapped, but not
filters. that can throw.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ca8d6057ae5cf611a4d921cdb6d0f34097bd8094