Rework `BlockHandle` to no longer have friend classes, and rework `ConvertToPersistent` so it fails if there are active outstanding pins (https://redirect.github.com/duckdb/duckdb/pull/14746)
This PR reworks the `BlockHandle` class to remove all of the friend
classes, and instead use a less error-prone interface that requires
passing the `BlockLock` into the methods when accessing members that
require locks (e.g. the buffer or memory usage fields).

In addition, we rework the `ConvertToPersistent` method so that it
throws an error when there are active readers, and fix cases where this
was the case. In all of the cases where this happened the handle would
have been destroyed shortly after the `ConvertToPersistent` was
completed anyway.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2d1b7d796d0e49e7a73ede14fd7d3033c1782626