[PySpark] Add more functions such as ascii, asin, btrim, char, corr, ... and fix differences in ordering of null values between PySpark and DuckDB (https://redirect.github.com/duckdb/duckdb/pull/14738)
For fixing the ordering of null values, I had to set
`default_null_order` to `nulls_first` for the whole DuckDB connection.
I've added the reasoning as comments in the code but let me know if I
should expand on that. Thanks to the PySpark tests in CI, we at least
now, that for the current test suite, this solves it.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4ce489734ee072c9205a18ddd4d5ab98550e88ce