Detect catalog changes on DROP IF EXISTS (https://redirect.github.com/duckdb/duckdb/pull/14742)
Fixes a bug where prepared statements of DROP IF EXISTS could run into
an `INTERNAL Error: Attempting to commit a transaction that is read-only
but has made changes - this should not be possible`. The problem is that
in case the table does not exist during bind, the catalog access is not
even registered as catalog READ access, and hence catalog change
detection cannot kick in (triggering rebind).

Upstream-commit: https://github.com/duckdb/duckdb/commit/1fe66c2ca4d6ae5aa88a46df6b9d4406e19ed35f