Internal https://redirect.github.com/duckdb/duckdb/pull/3263: Window Distinct Deadlock (https://redirect.github.com/duckdb/duckdb/pull/14775)
More like a hang. Seems to be caused by a local state being allocated by
the pipeline code during a DISTINCT finalize step. This would increment
the thread count used by the DISTINCT aggregator without creating a task
to process some rows.

The fix is to not use the thread number but actually count the number of
local sorts created.

fixes: duckdb-labs/duckdb-internalhttps://redirect.github.com/duckdb/duckdb/pull/3263
Upstream-commit: https://github.com/duckdb/duckdb/commit/dd01b20840f687936de23530a45bfee36aa5a3b1