Require `capacity` in ValidityMask (https://redirect.github.com/duckdb/duckdb/pull/14846)
This PR modifies the `ValidityMask` so it requires a capacity. The
capacity can then be used to verify that accesses to the validity mask
are in range. Note that this was already partially present
(`target_count`). In debug mode we verify that all accesses to the
validity mask are `< capacity`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b6ca9af3ed8fd49a914237ad3368bce248034b9e