return InsertionOrderPreservingMap from TableFunction to_string (https://redirect.github.com/duckdb/duckdb/pull/14835)
Follow up of https://github.com/duckdb/duckdb/pull/13109.

This PR extends the aforementioned by allowing the
TableFunction::to_string method to also return a map instead of a raw
string. This allows TableFunctions to return structured information for
the explain output.

Additionally, this PR reworks the parameters to the `to_string` function
to also take a reference to the table function. This is required for the
new delta attach functionality to be able to deduce which table is being
scanned.

With these two things the delta extension can make its explain output
for scanning attached tables a lot prettier by showing the name of the
table scanned (thanks @djouallah for this suggestion
https://github.com/duckdb/duckdb_delta/issues/117):

```
┌─────────────────────────────┐
│┌───────────────────────────┐│
││       Physical Plan       ││
│└───────────────────────────┘│
└─────────────────────────────┘
┌───────────────────────────┐
│           UNION           ├──────────────┐
└─────────────┬─────────────┘              │
┌─────────────┴─────────────┐┌─────────────┴─────────────┐
│        DELTA_SCAN         ││        DELTA_SCAN         │
│    ────────────────────   ││    ────────────────────   │
│         Table: dt1        ││         Table: dt2        │
│    Projections: c_name    ││    Projections: c_name    │
│                           ││                           │
│        ~150000 Rows       ││        ~150000 Rows       │
└───────────────────────────┘└───────────────────────────┘
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/24680c5a73c491d9d96920a31440a2912c2f2831