[Python] Fix various issues uncovered by https://redirect.github.com/duckdb/duckdb/pull/12959 (https://redirect.github.com/duckdb/duckdb/pull/13149)
This PR references https://redirect.github.com/duckdb/duckdb/pull/12959

Namely points 1 and 5 have not been addressed by this PR.
Both of which are not trivially fixed.
1 because of the very overloaded responsibilities of the createDataFrame
method.
5 because this is a problem at the Relation level.

Fixes are as follows:
- ColumnExpression can now be created with multiple names, to qualify
the column with a table.
- (pyspark) DataFrame methods have been adjusted to return a
non-ambiguous Column
- DuckDBPyRelation::Join has been adjusted to turn a list of strings
into a USING clause
- `desc` and `asc` have been added to `spark.sql.functions`

Relation's constructor has been adjusted to take an `alias` parameter,
this is now a private member of the Relation and is `const`.
Removing the need for `GetAlias`. The method has not been touched by
this PR however.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6cb7cec64751d9ede8919f98afbdfd5a663cae9b