Add `struct_concat` (https://redirect.github.com/duckdb/duckdb/pull/14853)
This PR adds a new `struct_concat` function to merge multiple structs
horizontally.
Concatenating structs with duplicate keys results in an error.
Concatenating unnamed structs is supported, but you cannot mix named and
unnamed.

A future cool improvement to consider when there are structs containing
keys of the same name would be to merge their entries into a new list of
the type of the duplicate key (or max logical type). The resulting new
member could even be a ARRAY as the number of duplicates would be known
at bind time.

Upstream-commit: https://github.com/duckdb/duckdb/commit/8ca81bab81b00bc8d97a228d57fada526e8f3be7