improve ReadAheadBuffer::AddReadHead error message (https://redirect.github.com/duckdb/duckdb/pull/14940)
We had a user running into this error when dealing with a corrupted
parquet file. It would be more helpful if ddb provided a bit more
information on which file was corrupted etc. I made this change to
provide more context on what went wrong in order to help
users/developers debug this issue more effectively.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1a3d614f0eec5a2198af8ba4ea06eb9adee9d5f8