Logical Sample requires child to have separate join order optimization (https://redirect.github.com/duckdb/duckdb/pull/14969)
Partially addresses
https://github.com/duckdblabs/duckdb-internal/issues/1470.
This enforces that if you have a join, a filter, and a sample, the join
is executed, then a sample is collected, and the sample is filtered on.
The error was that relations could be re-ordered around the sample, and
in effect push filters down below a join.

I still think the SAMPLE is applied in the correct place semantically.
If the sample is applied after filters and group by's to match syntax to
semantics, then it has the same effect as a limit. If someone wants a
sample of a result, they can just wrap their query in another `SELECT *
FROM {query} USING SAMPLE(1)`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/47490b7f4fb806232077046821e20e9792a7c180